pyproject.toml
[project]
name = "prospector-jobs"
version = "0.1.0"
description = "Agent that scans job boards to find companies hiring CPO/CPTO — potential fractional product leadership clients"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
browser = ["playwright>=1.40"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
]

[project.scripts]
prospector = "prospector_jobs.main:main"

[tool.setuptools.packages.find]
include = ["prospector_jobs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
